% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_metrics.R
\name{dep_analysis}
\alias{dep_analysis}
\title{Analysis: differential expression}
\usage{
dep_analysis(expr, group)
}
\arguments{
\item{expr}{A expression table file (at peptide level)}

\item{group}{The grouping info}
}
\description{
Analysis: differential expression
}
