% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_cor.R
\name{DEPanalysis}
\alias{DEPanalysis}
\title{Calculating RC value; Plotting a scatterplot}
\usage{
DEPanalysis(expr, group)
}
\arguments{
\item{expr_dt}{A expression table file (at peptide level)}

\item{meta_dt}{A metadata file}

\item{output_dir}{A directory of the output file(s)}
}
\description{
Calculating RC value; Plotting a scatterplot
}
