% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_protein_report}
\alias{generate_protein_report}
\title{Generate Quartet Proteomics report}
\usage{
generate_protein_report(
  qc_result,
  report_template,
  report_dir = NULL,
  report_name = NULL
)
}
\arguments{
\item{qc_result}{list}

\item{report_template}{character}

\item{report_dir}{character}

\item{report_name}{character}
}
\description{
Use calculated Met result to generate report
}
\examples{
# 加载示例数据
example_prot_data_path <- system.file("extdata", "proteomics_pipeline_data_template.csv", package = "protqc")
example_prot_metadata_path <- system.file("extdata", "proteomics_pipeline_meta_template.csv", package = "protqc")
qc_result <- qc_conclusion(example_prot_data_path, example_prot_metadata_path)

# 指定包内文档的路径
report_template <- system.file("extdata", "quartet_template.docx", package = "protqc")

# 运行函数
generate_protein_report(qc_result = qc_result, report_template = report_template)

}
