% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_report.R
\name{GenerateReport}
\alias{GenerateReport}
\title{Generate Quartet Proteomics report}
\usage{
GenerateReport(Prot_result = NULL, doc_file_path = NULL, output_path = NULL)
}
\arguments{
\item{Prot_result}{list}

\item{output_path}{character}

\item{temp_doc_path}{character}
}
\value{
word file
}
\description{
Use calculated Met result to generate report
}
\examples{
# 加载示例 Protein_result 对象
prot_result_path <- system.file("extdata", "Protein_result_example.RData", package = "protqc")
load(prot_result_path)

# 指定包内文档的路径
doc_file_path_example <- system.file("extdata", "Quartet_temp.docx", package = "protqc")

# 假设输出路径是当前工作目录
output_path_example <- getwd()

# 运行函数
GenerateReport(Prot_result=prot_result, doc_file_path=doc_file_path_example, output_path=output_path_example)


}
