% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_data_genesymbols.R
\docType{data}
\name{historical_data_genesymbols}
\alias{historical_data_genesymbols}
\title{Data of historical data sets at protein levels.}
\format{
A data frame with 9247 rows and 385 variables
}
\usage{
historical_data_genesymbols
}
\description{
Data of historical data sets at protein levels.
}
\keyword{datasets}
