% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_data_peptides.R
\docType{data}
\name{historical_data_peptides}
\alias{historical_data_peptides}
\title{Data of historical data sets at peptide levels.}
\format{
A data frame with 101158 rows and 73 variables
}
\usage{
historical_data_peptides
}
\description{
Data of historical data sets at peptide levels.
}
\keyword{datasets}
