% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_qc.R
\docType{data}
\name{historical_qc}
\alias{historical_qc}
\title{The performances for historical data sets.}
\format{
A data frame with 32 rows and 7 variables
}
\usage{
historical_qc
}
\description{
The performances for historical data sets.
}
\keyword{datasets}
