% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_qc_norm.R
\docType{data}
\name{historical_qc_norm}
\alias{historical_qc_norm}
\title{The performances for historical data sets (Normalized).}
\format{
A data frame with 32 rows and 9 variables
}
\usage{
historical_qc_norm
}
\description{
The performances for historical data sets (Normalized).
}
\keyword{datasets}
