% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_qc_stat.R
\docType{data}
\name{historical_qc_stat}
\alias{historical_qc_stat}
\title{Statistics of the performances for historical data sets.}
\format{
A data frame with 7 rows and 2 variables
}
\usage{
historical_qc_stat
}
\description{
Statistics of the performances for historical data sets.
}
\keyword{datasets}
