% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{input_data}
\alias{input_data}
\title{Input files}
\usage{
input_data(pro_path, meta_path, pep_path = NULL)
}
\arguments{
\item{pro_path}{A file path of the expression table file (at protein level)}

\item{meta_path}{A file path of the metadata file}

\item{pep_path}{A file path of the expression table file (at peptide level)}
}
\description{
Input files
}
