% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{plot_corr}
\alias{plot_corr}
\title{Plotting a scatterplot}
\usage{
plot_corr(expr_dt_path, meta_dt_path, output_dir)
}
\arguments{
\item{expr_dt_path}{A file path of the expression table file}

\item{meta_dt_path}{A file path of the metadata file}

\item{output_dir}{A directory of the output file(s)}
}
\description{
Plotting a scatterplot
}
