% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protqc-package.R
\docType{package}
\name{protqc-package}
\alias{protqc}
\alias{protqc-package}
\title{protqc: protqc}
\description{
Quality assessment of proteomic data for Quartet Project.
}
\author{
\strong{Maintainer}: Qiaochu Chen \email{chenqiaochu@sjtu.edu.cn}

Authors:
\itemize{
  \item Qiaochu Chen \email{chenqiaochu@sjtu.edu.cn}
}

}
\keyword{internal}
