% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pipelines.R
\name{qc_allmetrics}
\alias{qc_allmetrics}
\title{Calculating all QC metrics}
\usage{
qc_allmetrics(pro_dt, meta_dt, pep_dt = NULL, output_dir = NULL)
}
\arguments{
\item{pro_dt}{A expression table file (at protein level)}

\item{meta_dt}{A metadata file}

\item{pep_dt}{A expression table file (at peptide level)}

\item{output_dir}{A directory for results}
}
\description{
Calculating all QC metrics
}
