% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_linear_norm.R
\name{qc_linear_norm}
\alias{qc_linear_norm}
\title{Linear normalization: 1~10}
\usage{
qc_linear_norm(x, x_min, x_max, decreasing = F)
}
\arguments{
\item{x}{A value to be normalized}

\item{x_min}{Queried minimum in historical values}

\item{x_max}{Queried maximum in historical values}
}
\description{
Linear normalization: 1~10
}
