% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pipelines.R
\name{qc_performance}
\alias{qc_performance}
\title{Label the performance: bad, fair, good, great}
\usage{
qc_performance(x, x_ref, cutoff = c(0, 0.2, 0.5, 0.8, 1))
}
\arguments{
\item{x}{The value of the metric to be labeled}

\item{x_ref}{The reference values of the metric}

\item{cutoff}{The cut-offs of the labeling}
}
\description{
Label the performance: bad, fair, good, great
}
