% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pipelines.R
\name{qc_rank}
\alias{qc_rank}
\title{Rank in historical performances}
\usage{
qc_rank(x, x_hist)
}
\arguments{
\item{x}{The value of the metric to be labeled}

\item{x_hist}{The reference values of the metric}
}
\description{
Rank in historical performances
}
