% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_metrics.R
\name{qc_recall}
\alias{qc_recall}
\title{Calculating Recall}
\usage{
qc_recall(expr_dt, meta_dt)
}
\arguments{
\item{expr_dt}{A expression table file (at peptide level)}

\item{meta_dt}{A metadata file (to map library to sample type)}
}
\description{
Uses the supplied text's majority-replicate micro Recall.
A reference peptide must be detected in floor(n / 2) + 1 replicates.
Overall Recall is total detected reference peptide-sample pairs divided
by total reference peptide-sample pairs for the sample groups present.
Reference pairs are deduplicated; missing expression peptides count as undetected.
}
