% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_metrics.R
\name{qc_snr}
\alias{qc_snr}
\title{Calculating SNR value; Plotting a PCA panel}
\usage{
qc_snr(expr_dt, meta_dt, output_dir = NULL)
}
\arguments{
\item{expr_dt}{A expression profile (at protein level)}

\item{meta_dt}{A metadata file}

\item{output_dir}{A directory of the output file(s)}
}
\description{
Calculating SNR value; Plotting a PCA panel
}
