% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_dataset.R
\docType{data}
\name{reference_dataset}
\alias{reference_dataset}
\title{Data of historical data sets at peptide levels.}
\format{
A data frame with 4865 rows and 4 variables
}
\usage{
reference_dataset
}
\description{
Data of historical data sets at peptide levels.
}
\keyword{datasets}
