% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclusion.R
\name{table_conclusion}
\alias{table_conclusion}
\title{Generating a table of conclusion}
\usage{
table_conclusion(pro_path, meta_path, output_dir = NULL, pep_path = NULL)
}
\arguments{
\item{pro_path}{A file path of the expression table file (at protein level)}

\item{meta_path}{A file path of the metadata file}

\item{output_dir}{A directory of the output file(s)}

\item{pep_path}{A file path of the expression table file (at peptide level)}
}
\description{
Generating a table of conclusion
}
