% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input_data}
\alias{input_data}
\title{Input files}
\usage{
input_data(exp_path, meta_path)
}
\arguments{
\item{exp_path}{A file path of the expression table file}

\item{meta_path}{A file path of the metadata file}
}
\description{
Input files
}
