% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{qc_conclusion}
\alias{qc_conclusion}
\title{Generating a table of conclusion}
\usage{
qc_conclusion(exp_path, meta_path, output_dir = NULL)
}
\arguments{
\item{exp_path}{A file path of the expression table file}

\item{meta_path}{A file path of the metadata file}

\item{output_dir}{A directory of the output file(s)}
}
\description{
Generating a table of conclusion
}
