% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_metrics.R
\name{qc_info}
\alias{qc_info}
\title{Statistics for basic information}
\usage{
qc_info(expr_dt, meta_dt)
}
\arguments{
\item{expr_dt}{A expression profile}

\item{meta_dt}{A metadata file}
}
\description{
Statistics for basic information
}
