% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclusion.R
\name{table_conclusion}
\alias{table_conclusion}
\title{Generating a table of conclusion}
\usage{
table_conclusion(exp_path, meta_path, output_dir = NULL)
}
\arguments{
\item{exp_path}{A file path of the expression table file}

\item{meta_path}{A file path of the metadata file}

\item{output_dir}{A directory of the output file(s)}
}
\description{
Generating a table of conclusion
}
